MATX-198 - Extracted code common between MaterialXSurfaceOverride and MaterialXTextureOverride to MaterialXShadingNodeImpl. (#477)
* MATX-198 - Extracted code common between MaterialXSurfaceOverride and MaterialXTextureOverride to MaterialXShadingNodeImpl.

* Removed TestFileNodeOverride.

* Minor comments.

* Refactoring MaterialXShadingNodeImpl.

* Cleaned up formatting and includes.
