+++ source/MaterialXContrib/MaterialXNode/Plugin.cpp
@@ -85,24 +85,6 @@ MStatus initializePlugin(MObject obj)
             &surfaceNodeClassification));
     }
 
-    ///////////////////////////////////////////////////////////////////////
-    {
-        const MString UserClassify("texture/2d:drawdb/shader/texture/2d/testFileTexture");
-
-        CHECK_MSTATUS(plugin.registerNode(
-            "testFileTexture",
-            TestFileNode::id,
-            TestFileNode::creator,
-            TestFileNode::initialize,
-            MPxNode::kDependNode,
-            &UserClassify));
-
-        CHECK_MSTATUS(MHWRender::MDrawRegistry::registerShadingNodeOverrideCreator(
-            "drawdb/shader/texture/2d/testFileTexture",
-            sRegistrantId,
-            TestFileNodeOverride::creator));
-    }
-
     return MS::kSuccess;
 }
 
@@ -126,14 +108,6 @@ MStatus uninitializePlugin(MObject obj)
         MHWRender::MDrawRegistry::deregisterShadingNodeOverrideCreator(
         MaterialXSurfaceOverride::DRAW_CLASSIFICATION,
         MaterialXSurfaceOverride::REGISTRANT_ID));
-
-    ///////////////////////////////////////////
-    CHECK_MSTATUS(plugin.deregisterNode(TestFileNode::id));
-    CHECK_MSTATUS(
-        MHWRender::MDrawRegistry::deregisterShadingNodeOverrideCreator(
-            "drawdb/shader/texture/2d/testFileTexture",
-            sRegistrantId));
-
 
 	return MS::kSuccess;
 }
